FIXED: Issue when looting Currency items were not added to currency tab after looting. FIXED: Bonus from spell power should work properly now. FIXED: Juggernaut proc(procing is in DB, removing in core) UPDATED: spell_proc table(reapply 2780_spell_proc.sql!)

+++ src/arcemu-world/Item.cpp
@@ -679,7 +679,7 @@ int32 Item::AddEnchantment( EnchantEntry* Enchantment, uint32 Duration, bool Per
 		}
 	
 		/* Only apply the enchantment bonus if we're equipped */
-		uint8 slot = m_owner->GetItemInterface()->GetInventorySlotByGuid( GetGUID() );
+		uint16 slot = m_owner->GetItemInterface()->GetInventorySlotByGuid( GetGUID() );
 		if( slot >= EQUIPMENT_SLOT_START && slot < EQUIPMENT_SLOT_END )
             ApplyEnchantmentBonus( Slot, APPLY );
 	}
